Unpacked data and updated worklog
Test data is successfully input into the program and stored into
a general dataframe. Each column of the dataframe has been separated
out into a data series for processing. Processing using python 2.7
will need to be outlined in a flow chart to lay out the ideal way to
approach this problem.