Add changes to main.py and pigeon.py
Added the method for parsing trial information: parseTrialInfo. Also
changed the constructor to convert pixels to distance when they are
stored, rather than have a separate method for doing so. Added info
to the test.xls testing spreadsheet, and also have successful conversion
from pixels to distance in the program.